% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{build_export_url}
\alias{build_export_url}
\title{Construct the Microsoft Graph exportSurveys URL for a given mode.}
\usage{
build_export_url(mode, experience_name, survey_uuid = NULL, cycle_id = NULL)
}
\description{
Three URL shapes, all rooted at \link{glint_graph_base}:
\itemize{
\item \code{cycle}: \code{.../experiences/{exp}/surveys/{uuid}/surveyCycles/{cid}/exportSurveys}
\item \code{survey}: \code{.../experiences/{exp}/surveys/{uuid}/exportSurveys}
\item \code{daterange}: \code{.../experiences/{exp}/exportSurveys}
}
}
\keyword{internal}
