% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{get_all_reports}
\alias{get_all_reports}
\title{Get All Reports for a Manager}
\usage{
get_all_reports(manager_id, data, emp_id_col, manager_id_col)
}
\arguments{
\item{manager_id}{Character string of manager's employee ID}

\item{data}{Data frame with survey data}

\item{emp_id_col}{Character string specifying the employee ID column name}

\item{manager_id_col}{Character string specifying the manager ID column name}
}
\value{
Character vector of employee IDs for all reports
}
\description{
Internal function to get all direct and indirect reports for a manager.
}
\keyword{internal}
