% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{glint_env}
\alias{glint_env}
\title{Read a required Glint env var, stopping with a helpful message if missing.}
\usage{
glint_env(var_name, label)
}
\description{
Read a required Glint env var, stopping with a helpful message if missing.
}
\keyword{internal}
