% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{get_all_reports}
\alias{get_all_reports}
\title{Get All Reports for a Manager}
\usage{
get_all_reports(manager_id, data)
}
\arguments{
\item{manager_id}{Character string of manager's employee ID}

\item{data}{Data frame with survey data including EMP ID and Manager ID}
}
\value{
Character vector of employee IDs for all reports
}
\description{
Internal function to get all direct and indirect reports for a manager.
}
\keyword{internal}
