% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivaglint-package.R
\docType{package}
\name{vivaglint-package}
\alias{vivaglint}
\alias{vivaglint-package}
\title{vivaglint: Analysis Tools for Viva Glint Survey Data}
\description{
Provides functions for importing, validating, and analyzing Viva Glint survey data exports. Includes tools for data reshaping, question-level analysis, multi-cycle comparisons, and organizational hierarchy analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/microsoft/vivaglint}
  \item Report bugs at \url{https://github.com/microsoft/vivaglint/issues}
}

}
\author{
\strong{Maintainer}: Eric Knudsen \email{ericknudsen@microsoft.com}

}
\keyword{internal}
